64 Refactor timezone select into component.
