+++ client/src/components/shared/components/TimezoneSelect/assets/timezones.json
